Split domains into packages
